+++ services/server/entity/src/content_profile_update_model.rs
@@ -13,11 +13,11 @@ pub struct Model {
     pub name: Option<String>,
 
     // Avatar image digest (no FK — resolved via query)
-    pub avatar_digest_type: Option<i8>,
+    pub avatar_digest_type: Option<i16>,
     pub avatar_digest_bytes: Option<Vec<u8>>,
 
     // Banner image digest (no FK — resolved via query)
-    pub banner_digest_type: Option<i8>,
+    pub banner_digest_type: Option<i16>,
     pub banner_digest_bytes: Option<Vec<u8>>,
 }
